universe_agent/requirements.txt
# Core LangChain and LLM
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20

# Vector Store
chromadb>=0.4.22

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
playwright>=1.41.0
lxml>=5.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
rich>=13.7.0
typer>=0.9.0

# Optional: Additional search tools
googlesearch-python>=1.2.3
tweepy>=4.14.0
